Refactors `process_geometry_targets()` to fill=1
Remove multiple dispatch, data typing,

+++ src/clustering/clustering.jl
@@ -16,7 +16,7 @@ end
     )::Raster{Int64, 2}
     apply_kmeans_clustering(
         raster::Raster{Float64, 2}, k::Int8; tol::Float64=1.0
-    )::Raster{Int64, 2}
+    )::Raster{Int, 2}
 
 Cluster targets sites by applying k-means to target (non-zero) cells in a raster.
 - Float64 raster is assumed to contain disturbance values, addressed buy 3d clustering
